Rename for better code-style
