+++ src/Components/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './Header/Header';
-import Footer from './Footer/Footer';
+import Header from './Header';
+import Footer from './Footer';
 import AddItem from '../Containers/AddItem';
 import VisibleTodoList from '../Containers/VisibleTodoList';
