+++ src/Components/Footer/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import FilterLink from '../../Containers/FilterLink';
 import { VisibilityFilters } from '../../Actions/index';
-import styles from './Footer.module.css';
+import styles from './style.module.css';
 
 const Footer = () => (
   <div className={styles.footer}>
