+++ src/Components/Header/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import style from './Header.module.css';
+import style from './style.module.css';
 
 class Header extends React.Component {
   getTime() {
