+++ src/Components/Link/index.jsx
@@ -1,14 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import styles from './style.module.css';
 
 const Link = ({ active, children, onClick }) => (
   <button
+    className={styles.link}
     type="button"
     onClick={onClick}
     disabled={active}
-    style={{
-      marginLeft: '4px',
-    }}
   >
     {children}
   </button>
