+++ src/Containers/FilterLink.jsx
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { setVisibilityFilter } from '../Actions/index';
-import Link from '../Components/Link/Link';
+import Link from '../Components/Link';
 
 const mapStateToProps = (state, ownProps) => ({
   active: ownProps.filter === state.visibilityFilter,
